Add files via upload
11 lesson new version

+++ Lesson11/src/main/java/org/courses/lesson11/controller/UserController.java
@@ -1,6 +1,9 @@
 package org.courses.lesson11.controller;
 
 import org.courses.lesson11.dto.User;
+import org.courses.lesson11.exception.NoRightsToChangeDatabaseException;
+import org.courses.lesson11.exception.TryToChangeDefaultUserException;
+import org.courses.lesson11.exception.UnableToSaveUserException;
 import org.courses.lesson11.service.UserService;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -8,6 +11,8 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
+
+import javax.servlet.http.HttpSession;
 
 @Controller
 public class UserController {
@@ -26,7 +31,10 @@ public class UserController {
     }
 
     @PostMapping("/add")
-    public String addUser(@ModelAttribute("user") User user) {
+    public String addUser(@ModelAttribute("user") User user, HttpSession session)
+            throws NoRightsToChangeDatabaseException, UnableToSaveUserException {
+        long currentUserId = (Long) session.getAttribute("id");
+        userService.checkIfCurrentUserCanChangeDatabase(currentUserId);
         userService.create(user);
         return "redirect:/home";
     }
@@ -39,27 +47,35 @@ public class UserController {
 
     @GetMapping("/update/{id}")
     public String updateUserGetMapping(@PathVariable("id") long id, Model model) {
-        if(userService.find(id).isPresent()){
+        if (userService.find(id).isPresent()) {
             User userToUpdate = userService.find(id).get();
             model.addAttribute("user", userToUpdate);
             return "edit";
-        }else {
-            throw new IllegalArgumentException("Invalid id: "+ id);
+        } else {
+            throw new IllegalArgumentException("Invalid id: " + id);
         }
     }
 
     @PostMapping("/update/{id}")
-    public String updateUser(@PathVariable("id") long id, @ModelAttribute("user") User user) {
+    public String updateUser(@PathVariable("id") long id,
+                             @ModelAttribute("user") User user, HttpSession session)
+            throws NoRightsToChangeDatabaseException, TryToChangeDefaultUserException, UnableToSaveUserException {
+        long currentUserId = (Long) session.getAttribute("id");
+        userService.checkIfCurrentUserCanChangeDatabase(currentUserId);
         user.setId(id);
         userService.update(user);
         return "redirect:/home";
     }
 
     @GetMapping("/delete/{username}")
-    public String deleteUser(@PathVariable("username") String username) {
-        userService.delete(userService.find(username).get());
+    public String deleteUser(@PathVariable("username") String username, HttpSession session)
+            throws NoRightsToChangeDatabaseException, TryToChangeDefaultUserException {
+        long currentUserId = (Long) session.getAttribute("id");
+        userService.checkIfCurrentUserCanChangeDatabase(currentUserId);
+        if (userService.find(username).isPresent()) {
+            userService.delete(userService.find(username).get());
+        }
         return "redirect:/home";
     }
-
 
 }
